test(content): isolate predecessor observation
